Initial commit
This particular was a regression type of problem which i solved with keras regressor with some data preprocessing with the accuracy of 73% on online score. Stilll working in tuning the data with Deep learning and creating new features with feature engineering and also trying to use different regressor for more accuracy.